FROMGIT: drivers: thermal: tsens: Add watchdog support
TSENS IP v2.3 onwards adds support for a watchdog to detect if the TSENS
HW FSM is stuck. Add support to detect and restart the FSM in the
driver. The watchdog is configured by the bootloader, we just enable the
watchdog bark as a debug feature in the kernel.

Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Link: https://lore.kernel.org/r/a314747664a065db592ad77da7beae68128a5b6e.1584015867.git.amit.kucheria@linaro.org
(cherry picked from commit d22066c1afcda84306c96483357b8477b088ed30
 git://git.kernel.org/pub/scm/linux/kernel/git/thermal/linux.git thermal/linux-next)

BUG=b:149784651
TEST=build for trogdor

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I28b050daeb84ab798c7376017fe9ba034a80ffdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2052820
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>