FROMGIT: media: venus: hfi_cmds.h: Replace zero-length array with flexible-array member
The current codebase makes use of the zero-length array language
extension to the C90 standard, but the preferred mechanism to declare
variable-length types such as these ones is a flexible array member[1][2],
introduced in C99:

struct foo {
        int stuff;
        struct boo array[];
};

By making use of the mechanism above, we will get a compiler warning
in case the flexible array does not occur last in the structure, which
will help us prevent some kind of undefined behavior bugs from being
inadvertently introduced[3] to the codebase from now on.

Also, notice that, dynamic memory allocations won't be affected by
this change:

"Flexible array members have incomplete type, and so the sizeof operator
may not be applied. As a quirk of the original implementation of
zero-length arrays, sizeof evaluates to zero."[1]

This issue was found with the help of Coccinelle.

[1] https://gcc.gnu.org/onlinedocs/gcc/Zero-Length.html
[2] https://github.com/KSPP/linux/issues/21
[3] commit 76497732932f ("cxgb3/l2t: Fix undefined behaviour")

Signed-off-by: Gustavo A. R. Silva <gustavo@embeddedor.com>
Signed-off-by: Stanimir Varbanov <stanimir.varbanov@linaro.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 380f3bbd9562dc93be2e3cadc329b15284fbedae
 git://linuxtv.org/svarbanov/media_tree.git venus-for-next-v5.8-part2)

BUG=b:149955289
TEST=video decoding still works on trogdor:
  watch -n .1 "cat /sys/kernel/debug/clk/*venus*/clk_rate"
  play youtube video
    => clock rates go up

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: Ie7c263241397138c3812d669be14a106f19267ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2172663
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>