BACKPORT: FROMLIST: Bluetooth: Add mgmt op set_wake_capable
When the system is suspended, only some connected Bluetooth devices
cause user input that should wake the system (mostly HID devices). Add
a list to keep track of devices that can wake the system and add
a management API to let userspace tell the kernel whether a device is
wake capable or not. For LE devices, the wakeable property is added to
the connection parameter and can only be modified after calling
add_device.

Signed-off-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
(am from https://patchwork.kernel.org/patch/11432029/)

Fixups:
  - Modified mgmt op number and array locations for chromium specific
    values

BUG=b:146927113
TEST=Built

Signed-off-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Change-Id: If9002764858b23ce2b195e6bfffb6b00ba308ea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2119432
Reviewed-by: Alain Michaud <alainm@chromium.org>