CHROMIUM: platform/chrome: Init extcon-tcss from EC LPC driver
Adding a DT entry with just a compat string, that too via ACPI, is a
round about way to register a platform device. Instead, just make
the extcon-tcss-cros-ec device a platform device child of the
cros-ec-lpc device. For situations where the extcon device is not
compiled in, this should just throw an error print and continue.

BUG=b:147758849, b:152430275
BUG=b:157252781
BUG=b:159083038
BUG=b:160793940
TEST=emerge-volteer chromeos-kernel-5_4; ./update_kernel.sh <DUT_IP>
Verify that the extcon device probes on volteer.
TOREVERT=Revert this patch when we switch to cros-ec-typec for mux
support.

TEST=Check that Thunderbolt docks enumerate as expected using ToT
firmare images on volteer.
TOREVERT= This is a workaround to accomodate the existence of the
temporary extcon driver. It should be reverted when chromeos-5.4
switches to using the Type C connector class driver.

[[rebase510(gwendal): Squash
ae0953da87e0 ("CHROMIUM: platform/chrome: lpc: Limit extcon to volteerteer")
85b422aaca42 ("CHROMIUM: platform/chrome: lpc: Optimize extcon Squashupport")
3cda3f6a543d ("CHROMIUM: platform/chrome: lpc: Add extcon support"),
fixing same area]

Signed-off-by: Prashant Malani <pmalani@chromium.org>
Change-Id: I64879b5271de077a72189292e7e3467b914bdf5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2121744
Reviewed-by: Alex Levin <levinale@chromium.org>