CHROMIUM: platform/chrome: Init extcon-tcss from EC LPC driver
Adding a DT entry with just a compat string, that too via ACPI, is a
round about way to register a platform device. Instead, just make
the extcon-tcss-cros-ec device a platform device child of the
cros-ec-lpc device. For situations where the extcon device is not
compiled in, this should just throw an error print and continue.

BUG=b:147758849, b:152430275
TEST=Check that Thunderbolt docks enumerate as expected using ToT
firmare images on volteer.
TOREVERT= This is a workaround to accomodate the existence of the
temporary extcon driver. It should be reverted when chromeos-5.4
switches to using the Type C connector class driver.

Signed-off-by: Prashant Malani <pmalani@chromium.org>
Change-Id: I64879b5271de077a72189292e7e3467b914bdf5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2121744
Reviewed-by: Alex Levin <levinale@chromium.org>