CHROMIUM: thunderbolt: Initialize TMU again on resume
The TMU will be reset after router exits sleep so in order to
re-configure it upon resume make sure the structure is initialized again
based on the current hardware state.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>

BUG=b:160191015
TEST=Observed no regressions with S0iX without an attached device.

TOREVERT=Once the upstream version of this patch lands in maintainer
tree, this patch should be reverted.

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: Id53b8726cfedd5394581c5fcc7b08d1477dc927d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2275610
Tested-by: Dana Alkattan <dana.alkattan@intel.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>