CHROMIUM: thunderbolt: Do not change default USB4 router notification timeout
Some early stage USB4 devices do not like that any of the enumerating
router config space fields (ROUTER_CS_1 - ROUTER_CS_4) are written after
the initial enumeration for example when entering sleep states. The
default timeout by the USB4 spec is 10 ms which should be fine for the
driver to handle.

For this reason do not change the notification timeout from the default
10 ms for USB4 routers.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>

BUG=b:160191015
TEST=Observed no regressions with S0iX without an attached device.

TOREVERT=Once the upstream version of this patch lands in maintainer
tree, this patch should be reverted.

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I05d4c5a3ca075669459bbd73008b55224f16014d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2275611
Tested-by: Dana Alkattan <dana.alkattan@intel.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>