CHROMIUM: soundwire: bus: use property to set interrupt masks
Add a slave-level property and program the SCP_INT1_MASK as desired by
the codec driver. Since there is no DisCo property this has to be an
implementation-specific firmware property or hard-coded in the driver.

The only functionality change is that implementation-defined
interrupts are no longer set for amplifiers - those interrupts are
typically for jack detection or acoustic event detection/hotwording.

Tested-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>
Reviewed-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>
Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>

Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:162008784
TEST=Test audio on volteer.
TOREVERT: Once upstream solution lands revert this.

Change-Id: I8d9329b8bd8f9c5cc1c27d9516619d3dbacb57e2
Signed-off-by: Mike Mason <michael.w.mason@intel.corp-partner.google.com>