UPSTREAM: mm: prevent a warning when casting void* -> enum
I recently build the RISC-V port with LLVM trunk, which has introduced a
new warning when casting from a pointer to an enum of a smaller size.
This patch simply casts to a long in the middle to stop the warning.  I'd
be surprised this is the only one in the kernel, but it's the only one I
saw.

(cherry picked from commit 4708f31885a0d3e41a03c39501df697e73406f6a)
casting void* -> enum"))

BUG=b:155225860
TEST=Build Mistral

Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Link: http://lkml.kernel.org/r/20200227211741.83165-1-palmer@dabbelt.com
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

Change-Id: Ic3aa81ebadaf3485ea3956efa5ac682d32c81af2
Signed-off-by: Kan Yan <kyan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2171783
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Kan Yan <kyan@chromium.org>
Reviewed-by: Srinivasa duvvuri <sduvvuri@google.com>
Tested-by: Kan Yan <kyan@chromium.org>
Commit-Queue: Kan Yan <kyan@chromium.org>