FROMGIT: remoteproc: sysmon: Inform current rproc about all active rprocs
Clients/services running on a remoteproc that booted up might need to be
aware of the state of already running remoteprocs. When a remoteproc boots
up (fresh or after recovery) it is not aware of the remoteprocs that booted
before it, i.e., the system state is incomplete. So to keep track of it we
send sysmon on behalf of all 'ONLINE' remoteprocs.

Acked-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Siddharth Gupta <sidgup@codeaurora.org>
Link: https://lore.kernel.org/r/1586389003-26675-4-git-send-email-sidgup@codeaurora.org
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 1f36ab3f6e3b791eadad94f792c874706e153b66
 https://git.kernel.org/pub/scm/linux/kernel/git/andersson/remoteproc.git/ rproc-next)

BUG=b:149924251
TEST=Boot trogdor with wifi/modem

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: I0d572cff807cabac47f02edbb997c05174a6916f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2169284
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>