FROMGIT: soc: qcom: rpmh-rsc: Remove get_tcs_of_type() abstraction
The get_tcs_of_type() function doesn't provide any value.  It's not
conceptually difficult to access a value in an array, even if that
value is in a structure and we want a pointer to the value.  Having
the function in there makes me feel like it's doing something fancier
like looping or searching.  Remove it.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Maulik Shah <mkshah@codeaurora.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Tested-by: Maulik Shah <mkshah@codeaurora.org>
Link: https://lore.kernel.org/r/20200413100321.v4.4.Ia348ade7c6ed1d0d952ff2245bc854e5834c8d9a@changeid
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 53d49fe1ff49196712c416ddb987634483423934
 git://git.kernel.org/pub/scm/linux/kernel/git/qcom/linux.git drivers-for-5.8)

BUG=b:149596107
TEST=See bug

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ia348ade7c6ed1d0d952ff2245bc854e5834c8d9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2138379
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>