Revert "FROMLIST: pinctrl: cherryview: Add quirk with custom translation of ACPI GPIO numbers"
The quirk breaks headset detection on Setzer board.
It is because Setzer has vendor HP instead of GOOGLE in the DMI.

Even if HP is added to the DMI matching field, there is an issue on
touchpad.

Revert that patch, and let the proper fix in vboot_reference fix the
crossystem issue.

This reverts commit 383d9bec92939d1d34c1097358590d75ddc87c2b.

BUG=b:143174998, b:153985719
TEST=check headphone detection works on setzer.

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2150577
Commit-Queue: Alex Levin <levinale@chromium.org>
Reviewed-by: Alex Levin <levinale@chromium.org>
Change-Id: I7a9607835ded65f49cafaf9c61d5a3da256fa61a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2153199