FROMGIT: platform/x86: intel_scu_ipc: Add managed function to register SCU IPC
Drivers such as intel_pmc_ipc.c can be unloaded as well so in order to
support those in this driver add a new function that can be called to
unregister the SCU IPC when it is not needed anymore.

We also add a managed version of the intel_scu_ipc_register() that takes
care of calling intel_scu_ipc_unregister() automatically when the driver
is unbound.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 7e18c89d6e37bfffa2b161b08f40b53d2ce7ee94
git://git.infradead.org/linux-platform-drivers-x86.git for-next)

BUG=b:151418064
TEST=No regressions

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I5321da866e974945709e9d9192b7293496d21254