FROMGIT: BACKPORT: platform/x86: intel_pmc_ipc: Convert to MFD
This driver only creates a bunch of platform devices sharing resources
belonging to the PMC device. This is pretty much what MFD subsystem is
for so move the driver there, renaming it to intel_pmc_bxt.c which
should be more clear what it is.

MFD subsystem provides nice helper APIs for subdevice creation so
convert the driver to use those. Unfortunately the ACPI device includes
separate resources for most of the subdevices so we cannot simply call
mfd_add_devices() to create all of them but instead we need to call it
separately for each device.

The new MFD driver continues to expose two sysfs attributes that allow
userspace to send IPC commands to the PMC/SCU to avoid breaking any
existing applications that may use these. Generally this is bad idea so
document this in the ABI documentation.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 25f1ca31e230598eaf3c38d387a355a64bd772a7
git://git.infradead.org/linux-platform-drivers-x86.git for-next)

Conflicts:
	drivers/platform/x86/Kconfig
	drivers/platform/x86/Makefile
	drivers/platform/x86/intel_pmc_ipc.c
	Due to
	commit ("b16482fd036f1330e8 FROMLIST: platform/x86: intel_pmc_ipc: Use XTAL freq based on cpuid")
	present on chromeos-5.4 kernel, but not on upstream.

BUG=b:151418064
TEST=No regressions

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I64b4f1af8701d8d2230b971b520e7b8fa0ed68c9