FROMGIT: blk-mq: In blk_mq_dispatch_rq_list() "no budget" is a reason to kick
In blk_mq_dispatch_rq_list(), if blk_mq_sched_needs_restart() returns
true and the driver returns BLK_STS_RESOURCE then we'll kick the
queue.  However, there's another case where we might need to kick it.
If we were unable to get budget we can be in much the same state as
when the driver returns BLK_STS_RESOURCE, so we should treat it the
same.

It should be noted that even if we add a whole bunch of extra kicking
to the queue in other patches this patch is still important.
Specifically any kicking that happened before we re-spliced leftover
requests into 'hctx->dispatch' wouldn't have found any work, so we
really need to make sure we kick ourselves after we've done the
splicing.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Ming Lei <ming.lei@redhat.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit ab3cee3762e5e69f27c302c43691289fdfc12316
 git://git.kernel.org/pub/scm/linux/kernel/git/axboe/linux-block.git for-5.8/block)

BUG=chromium:1061950
TEST=Reboot tests

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I1f95c459e51962b8d2c83e869913b6befda2255c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2125232
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>