UPSTREAM: drm/i915/selftests: Disable C-states when measuring RPS frequency response
Let's isolate the impact of cpu frequency selection on determing the GPU
throughput in response to selection of RPS frequencies.

For real systems, we do have to be concerned with the impact of
integrating c-states, p-states and rp-states, but for the sake of
proving whether or not RPS works, one baby step at a time.

For the record, as one would hope, it does not seem to impact on the
measured performance, but we do it anyway to reduce the number of
variables. Later, we can extend the testing to encourage the the
cpu/pkg to try and sleep while the GPU is busy.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200421142236.8614-1-chris@chris-wilson.co.uk
Link: https://patchwork.freedesktop.org/patch/msgid/20200421142236.8614-1-chris@chris-wilson.co.uk
(cherry picked from commit cf9ba27840c26f649a90bb04db5b455d47810292)

BUG=b:168669388
TEST=Check for GPU frequency samples above 1100MHz.
     - Login as guest and launch WebGL Aquarium with 30,000 fish.
     - Monitor frequency statistics.
       `socwatch -f cpu -f gfx -t 120`
       alternatively
       `while true; do cat /sys/class/drm/card0/gt_act_freq_mhz; sleep 0.1; done`

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: I92d3de8bd924187cbbcd2f6b40894603655bf0dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2472671
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
(cherry picked from commit 89060f35aa0ec85d70fadc9e265998dd24904c6d)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2518569
Reviewed-by: Alex Levin <levinale@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Alex Levin <levinale@google.com>