FROMGIT: arm64: cpufeature: Relax check for IESB support
We don't care if IESB is supported or not as we always set
SCTLR_ELx.IESB and, if it works, that's really great.

Relax the ID_AA64MMFR2.IESB cpufeature check so that we don't warn and
taint if it's mismatched.

[will: rewrote commit message]

Signed-off-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>
Reviewed-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Link: https://lore.kernel.org/r/20200421142922.18950-2-will@kernel.org
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit 9d3f888135505dda7d5539a5756440d298f08173
 git://git.kernel.org/pub/scm/linux/kernel/git/arm64/linux for-next/core)

BUG=b:144577102
TEST=No more sanity check errors

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I0498fae8e566cc51eb6667a45018bdaa70b193dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2204895
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>