CHROMIUM: arm64: dts: sc7180: Temporarily hack out CQE for trogdor
We're seeing I/O hangs with CQE.  Let's disable it so people can get
back to stability while we're investigating.

Technically it's possible to just do the lighter-weight hacking out of
just "dcmd" but this patch is nice because it only touches a file
that's not yet upstream and changes to it will be squashed when it's
actually sent upstream, so temporary hacks like this will just
"vanish".

BUG=b:153142831
TEST=login_LoginSuccess.stress

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ib08737d47c7fd4e97d560d220a25eede4dd08f2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2161585
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>