FROMGIT: arm64: cpufeature: Extend comment to describe absence of field info
When a feature register field is omitted from the description of the
register, the corresponding bits are treated as STRICT RES0, including
for KVM guests. This is subtly different to declaring the field as
HIDDEN/STRICT/EXACT/0, so update the comment to call this out.

Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit 433022b58e628418020155e1f6a3d59c06a474c0
 git://git.kernel.org/pub/scm/linux/kernel/git/arm64/linux for-next/cpufeature)

BUG=b:144577102
TEST=dmesg | grep 'CPU features:'

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I327c652be8a3c41c7ac0c2ce62561018f404f000
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2220344
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>