CHROMIUM: BACKPORT: FROMLIST: platform/x86: Add Intel Input Output Manager (IOM) driver
Input Output Manager (IOM) is part of the Tiger Lake SoC that
configures the Type-C Sub System (TCSS). IOM is a micro controller
that handles Type-C topology, configuration and PM functions of
various Type-C devices connected on the platform.

This driver helps read relevant information such as Type-C port
status (whether a device is connected to a Type-C port or not) and
the activity type on the Type-C ports (such as USB, Display Port,
Thunderbolt), for consumption by other drivers.

Currently intel_iom_port_status() API is exported by this driver,
that has information about the Type-C port status and port activity
type.

Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>

(am from https://patchwork.kernel.org/patch/11666447/)

Conflicts:
	drivers/platform/x86/Makefile

This patch addresses comments on the above patch, but v2 of this
patch will be posted after 5.9-rc1 is released (to be in line with
the expectations of the community around posting new patches).

BUG=b:156016218
TEST=None.
TOREVERT: Once the upstream patch lands in maintainer's branch, this
patch should be reverted.

Change-Id: If4ce2b05629dc8f5b8352ef027045e8fd1605297
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2317033
Reviewed-by: Alex Levin <levinale@google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>