CHROMIUM: dts: sc7180: Fix IRQ on headset jack detection
When there is no headphone plugged, IRQ will be in high level.
The interrupt pin from ALC5682 should be pulled up.

Interrupt should be triggered on both edges of IRQ.
Setting this type in dts will be neglected because codec driver rt5682.c
will override it. We should kept it correct, so modify it in dts.

BUG=b:147070839 b:154899488
TEST=test audio with whole patchset

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: Iffa6589778bad4e79ccaefa3998ce16831ba3c51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2294935
Reviewed-by: Douglas Anderson <dianders@chromium.org>