CHROMIUM: arm64: dts: qcom: Hack around eMMC errors on sc7180-trogdor-rev1
For unknown reasons the eMMC that is stuffed on sc7180-trogdor-rev1
doesn't work unless you do the "restore DLL config" bit that we used
to have to do on sdm845.

While this is being debugged, temporarily hack just this revision of
trogdor to think it's got an sdm845 sdhci controller.

This patch is expected to be reverted.

BUG=b:155826689
TEST=Boot up rev1

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I64e66fea53ce8893e84e95498dae020a15ed4f82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2194522
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Alexandru M Stan <amstan@chromium.org>
Tested-by: Alexandru M Stan <amstan@chromium.org>