CHROMIUM: Config: Enable CONFIG_SND_SOC_INTEL_SKYLAKE
This is partial revert for
commit b07903d2276abf50d978a88c0dd7fa4170aec861.

After discussion with Intel it was decided that for SOF we should
blacklist SKL modules for Volteer/Dedede/Deltaur in the Alsa overlay instead of
disabling those modules in splitconfig for all platforms.

BUG=b:152749177, b:151807500, b:155256301, b:157296071
TEST=check sound card on Eve
(lsmod | grep snd_soc_kbl_rt5663_rt5514_max98927)
TEST=check sound card on Volteer and Dedede(using SOF) to make sure it is
not affected.

Cq-Depend: chromium:2302750, chromium:2193851, chromium:2304449
Change-Id: Id02e3d46578e13463329f6bf34eb0e53316a2e23
Signed-off-by: Radoslaw Biernacki <biernacki@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2199299
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>