CHROMIUM: arm64: dts: qcom: Fix compat string for trogdor-rev0
The bootloader sees the "-sku0" (which is intended for trogdor-rev1+)
and matches that.  That's because the bootloader's preference is:

google,trogdor-rev0-sku0
google,trogdor-sku0
google,trogdor-rev0
google,trogdor

Let's be really explicit about the old revisions.  We'll match both
sku0 and sku1 even if I'm not sure anyone has a board that says sku1.

BUG=b:157234701
TEST=Build and boot on trogdor-rev0.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Iba0898c74096b7b698ae44eefc068ed05d85be92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2212542
Reviewed-by: Evan Green <evgreen@chromium.org>