CHROMIUM: ASoC: codecs: rt*-sdw.c: fix off-by-one port count
The variable num_of_port is currently initialized to 1, even when DP0
is not enabled. This results in a number of ports that's too large.

Also remove rt715 sink port handling. In rt715_read_prop(), the
prop->sink_ports mask is zero, so there's no point in trying to
allocated any resources for non-existent ports.

Cc: Oder Chiou <oder_chiou@realtek.com>
Cc: Shuming Fan <shumingf@realtek.com>
Cc: Jack Yu <jack.yu@realtek.com>

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>
Reviewed-by: Rander Wang <rander.wang@linux.intel.com>
Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>

Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:162008784
TEST=Test audio on volteer.
TOREVERT: Once upstream solution lands revert this.

Change-Id: I083c9574d70231cf7422cf625c8379313cf39f4c
Signed-off-by: Mike Mason <michael.w.mason@intel.corp-partner.google.com>